requirements.txt
Pillow
loguru
